Move Presenter and PresenterStore to Architecture folder.
