+++ android-sdk/src/main/java/com/checkout/android_sdk/Architecture/Presenter.kt
@@ -1,0 +1,5 @@
+package com.checkout.android_sdk.Architecture
+
+
+interface Presenter {
+}
