--- android-sdk/src/main/java/com/checkout/android_sdk/Presenter/Presenter.kt
@@ -1,5 +1,0 @@
-package com.checkout.android_sdk.Presenter
-
-
-interface Presenter {
-}
